feat(uwsgi-user-harakiri): apply rev proxy timeout
Because a request may stay in uWSGI socket backlog for uncertain
time, the reverse proxy may time out during or even before the
handling of the request, and uWSGI may still handle the request
with successful commitment. In order to avoid inconsistent result,
this feature sets the user harakiri timeout on a per-request basis
based on request arrival time and timeout values set by the reverse
proxy as WSGI environment variables.